+++ frontend/package.json
@@ -18,6 +18,7 @@
     "@types/react-router-dom": "^5.1.6",
     "@types/react-select": "^3.0.21",
     "@types/styled-components": "^5.1.3",
+    "awesome-debounce-promise": "^2.1.0",
     "axios": "^0.21.1",
     "canvg": "^3.0.7",
     "d3": "^6.2.0",
