+++ frontend/src/ui/Select/Select.tsx
@@ -1,6 +1,7 @@
 import { useTheme } from '@material-ui/core';
-import { type } from 'os';
 import React from 'react';
+
+import AwesomeDebouncePromise from 'awesome-debounce-promise';
 
 import ReactSelect, { Props } from 'react-select';
 import AsyncSelect from 'react-select/async';
@@ -14,28 +15,6 @@ export interface SelectProps<T> extends Props<T> {
   error?: string;
   disabled?: boolean;
 }
-
-const fetchMatchOccupations = async (input: string) => {
-  /*
-   * Fetch occupations matching the given input string via the soc-smart-list endpoint in the DjangoApiClient
-   */
-  console.log('Fetching occupations matching the input ' + input);
-  // TODO: Change the input.length to input complete
-  if (input.length >= 0) {
-    const client = new DjangoApiClient();
-    const occupations = await client.getOccupations(input);
-
-    // Sort by occupation SOC code if there is not input yet
-    if (input === '') {
-      occupations.sort((job1, job2) => job1.code.localeCompare(job2.code));
-    }
-
-    // Each occupation has the name and code attributes. These attributes are handled by OccupationSele
-    return occupations.map(function (occupation) {
-      return occupation;
-    });
-  }
-};
 
 export const Select = <T,>({
   options,
@@ -130,10 +109,11 @@ export interface OccupationSelectProps
 export const OccupationSelect = ({
   occupations,
   onSelectOccupation = () => {},
+  fetchOptions = (input: string) => {},
   ...rest
 }: OccupationSelectProps): JSX.Element => (
   <SelectAsync
-    loadOptions={fetchMatchOccupations}
+    loadOptions={fetchOptions}
     aria-label="occupation-select"
     options={occupations}
     placeholder={'Select occupation...'}
